+++ modules/daddy_tops/daddy_app.py
@@ -1,15 +1,7 @@
 import os
 from flask_sqlalchemy import SQLAlchemy
 from daddy_app import app
-
-db = SQLAlchemy(app)
-
-def initDB():
-    db.create_all()
-    user = User(username="admin")
-    user.hash_password(app.config['admin_password'])
-    db.session.add(user)
-    db.session.commit()
+from views import initDB
 
 if __name__ == '__main__':
     if not os.path.exists('db.sqlite'):
